% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.excel.R
\name{read.excel}
\alias{read.excel}
\title{A function to read the excel clipboard}
\usage{
read.excel(header = TRUE, ...)
}
\arguments{
\item{header}{Import header}
}
\description{
A function to read the excel clipboard
}
\examples{
data = read.excel()
read.excel()
}
\keyword{excel}
