% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.excel.R
\name{write.excel}
\alias{write.excel}
\title{A function to paste R dataframes to the excel clipboard}
\usage{
write.excel(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{}

\item{row.names}{}

\item{col.names}{}
}
\description{
A function to paste R dataframes to the excel clipboard
}
\examples{
data = write.excel(datos)
write.excel()
}
\keyword{excel}
